Refactors to avoid overshadowing
