+++ tests/download_pwv_data.py
@@ -29,7 +29,7 @@ from pwv_kpno._download_pwv_data import _download_data_for_year
 from pwv_kpno._download_pwv_data import _read_file
 from pwv_kpno._download_pwv_data import _suomi_date_to_timestamp
 from pwv_kpno.pwv_atm import _timestamp
-from pwv_kpno._settings import Settings
+from pwv_kpno._package_settings import Settings
 
 __authors__ = ['Daniel Perrefort']
 __copyright__ = 'Copyright 2017, Daniel Perrefort'
